Attempted to implement IPv6ChangedDialog
The changes in this branch have rendered the Nexus Dex-Trade module useless because it cannot properly load the API keys from storage.json even though the keys clearly exist (encrypted, of course, as they were before).  Because this is probably a problem in how the guard state governing the opening and closing of the dialog, the whole of the changes in this commit need to be analyzed and correct.